[runtime/iobuf] cleanup
