+++ runtime/Cargo.toml
@@ -71,6 +71,7 @@ arbitrary = [
 	"dep:arbitrary",
 	"dep:commonware-conformance",
 ]
+bench = []
 external = [ "pin-project" ]
 test-utils = []
 iouring = [ "iouring-network", "iouring-storage" ]
@@ -95,3 +96,4 @@ path = "src/utils/buffer/benches/bench.rs"
 name = "buffer_pool"
 harness = false
 path = "src/iobuf/benches/bench.rs"
+required-features = ["bench"]
