+++ runtime/src/benchmarks/mod.rs
@@ -1,6 +1,4 @@
 //! Utilities for running benchmarks.
 
 pub mod context;
-#[doc(hidden)]
-pub mod iobuf;
 commonware_macros::stability_mod!(BETA, pub mod tokio);
